#1 fixing class naming
